+++ packages/ak-inline-dialog/src/index.register.js
@@ -1,5 +1,0 @@
-import { define } from 'skatejs';
-import definition from './index.js';
-
-// Expose the WebComponents
-export default define('ak-inline-dialog', definition);