+++ packages/ak-inline-dialog/stories/dialog-remove-story.js
@@ -1,13 +1,13 @@
 import reactify from 'akutil-react';
 import { storiesOf } from '@kadira/storybook';
-import { define, vdom } from 'skatejs'; // eslint-disable-line no-unused-vars
+import { vdom } from 'skatejs'; // eslint-disable-line no-unused-vars
 import { name } from '../package.json';
-import definition from '../src/index';
+import AkInlineDialog from '../src/index';
 
-const { React, ReactDOM, uniqueWebComponent } = window;
+const { React, ReactDOM } = window;
 const { Component } = React;
 
-const Dialog = reactify(uniqueWebComponent(name, definition, define), {
+const Dialog = reactify(AkInlineDialog, {
   React,
   ReactDOM,
 });