+++ packages/ak-layer/CHANGELOG.md
@@ -1,5 +1,15 @@
-<a name="56.1.1"></a>
-## 56.1.1 (2016-07-14)
+<a name="56.1.2"></a>
+## 56.1.2 (2016-07-18)
+
+
+
+<a name="ak-layer@56.1.2"></a>
+## ak-layer@56.1.2 (2016-07-18)
+
+
+### Bug Fixes
+
+* **skate:** SkateJS update to beta.10 ([620034a](https://bitbucket.org/atlassian/atlaskit/commits/620034a))